requirements.txt
aurorapy==0.2.6
future==0.18.2
paho-mqtt==1.6.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
Shapely==1.8.1.post1
six==1.16.0
suntime==1.2.5
